+++ projects/frontend/es-app/src/guis/menus/GuiMenu.h
@@ -24,7 +24,7 @@ class GuiMenu : public GuiMenuBase
       UISettings,
       Sound,
       Network,
-      Scrapper,
+      Scraper,
       Advanced,
       Bios,
       License,