+++ projects/frontend/es-app/src/scraping/INotifyScrapeResult.h
@@ -17,7 +17,7 @@ class INotifyScrapeResult
     virtual void GameResult(int index, int total, FileData* result) = 0;
 
     /*!
-     * @brief Notify the caller of scrapping is being aborted immediately
+     * @brief Notify the caller of scraping is being aborted immediately
      */
     virtual void ScrapingComplete(ScrapeResult reason) = 0;
 };