+++ projects/frontend/es-app/src/scraping/scrapers/thegamedb/TheGameDBEngine.h
@@ -20,9 +20,9 @@ class TheGameDBEngine : public IScraperEngine
 
     /*!
      * @brief Run the scraper using the given methods, on the given system list and report progress using notifyTarget
-     * @param method Scrapping method
+     * @param method Scraping method
      * @param systemList System list
-     * @param notifyTarget Interface for reporting scrapping progression
+     * @param notifyTarget Interface for reporting scraping progression
      * @return True if everything has been successful. False if cancelled, quota reached or fatal error occurred
      */
     bool RunOn(ScrappingMethod method, const SystemManager::SystemList& systemList, INotifyScrapeResult* notifyTarget,
@@ -30,9 +30,9 @@ class TheGameDBEngine : public IScraperEngine
 
     /*!
      * @brief Run the scraper using the given methods, on the given single game and report progress using notifyTarget
-     * @param method Scrapping method
+     * @param method Scraping method
      * @param singleGame Single game to scrape
-     * @param notifyTarget Interface for reporting scrapping progression
+     * @param notifyTarget Interface for reporting scraping progression
      * @return True if everything has been successful. False if cancelled, quota reached or fatal error occurred
      */
     bool RunOn(ScrappingMethod method, FileData& singleGame,